fix: the workspace API pointed at namespaces that moved, and shadowed the panel
Four things, all found by walking every page of a panel that consumes this
package rather than by reading it.

WorkspaceController imported HasApiResponses, ResolvesWorkspace,
PaginatedCollection and WorkspaceResource from Mod\Api. Those live in
Core\Api — the same classes, renamed — so the class could not be loaded at
all and every route through it was a container error.

Routes/api.php then imported the controller itself from
Core\Mod\Api\Controllers\WorkspaceController, which exists nowhere. It is
this package's own, in Core\Tenant\Controllers.

Those routes also claimed a bare prefix('workspaces'). fireApiRoutes()
applies the api middleware group and deliberately no prefix, because API
routes are expected to be domain-scoped — so a bare path is claimed on
every domain the application serves. It shadowed the consuming panel's own
/workspaces and /workspaces/create: the workspace list answered with a
container error, and the creation wizard tried to load a workspace called
"create". Everything is under v1/ now, which is the convention the event's
own documentation gives as its example.

EntitlementService::getCurrentUsage declares int and returned whatever the
cache handed back — a string, on any store that serialises to text. It only
failed on the second read of a figure, which is why the usage page broke
and the dashboard beside it did not.

And account_deletion_requests had no migration. The model, the job, the
console command, both Livewire components and the mail template all ship;
the table never did, so /account/settings threw on every render — it has to
look for a pending request to know whether to show the cancel banner.

+++ Services/EntitlementService.php
@@ -1221,11 +1221,16 @@ class EntitlementService
 
         // Use tagged cache if available for O(1) invalidation
         if ($this->supportsCacheTags()) {
-            return Cache::tags($this->getWorkspaceCacheTags($workspace, 'usage'))
+            // Cast on the way out. A cache store that serialises to text —
+            // database, redis — returns a string, and the signature says int,
+            // so the second read of any usage figure was a TypeError. It only
+            // ever showed up once something had been cached, which is why the
+            // usage page failed and the dashboard beside it did not.
+            return (int) Cache::tags($this->getWorkspaceCacheTags($workspace, 'usage'))
                 ->remember($cacheKey, self::USAGE_CACHE_TTL, $callback);
         }
 
-        return Cache::remember($cacheKey, self::USAGE_CACHE_TTL, $callback);
+        return (int) Cache::remember($cacheKey, self::USAGE_CACHE_TTL, $callback);
     }
 
     /**
